- Restock table is now split up per profession - Restock table now shows how many vestiges other characters have to restock as well, with the vestige that is cheapest for them. - LST now stores, and syncs, which character is your crafter for each leggo, and what their professions are
Since this feature requires a lot of backend changes relying on data, you will need to log in on each crafter to update the data, and then sync it if you use multiple accounts.
LST should try and warn you when it's missing data, check your chat if items seem to be missing!